clean up the service structures
